Use patched version of NtProtectVirtualMemory instead of syscalls
VMProtect hooks NtProtectVirtualMemory and similar calls (like VirtualProtect) to always return ACCESS_DENIED in order to achieve read only state of the memory.

Syscalling NtProtectVirtualMemory to avoid the hooked version is fine for Windows, but since it relies on the layout of the DLL being consistent, it doesn't work on Linux/Wine.

Since the hook is effectively placed at the beginning of the NtProtectVirtualMemory, we can take the bytes from the original beginning of the function and replace the hook with those bytes during runtime of the game.

Big thanks to mdias for the original implementation of this idea in RS_ASIO.